requirements.txt
asgiref==3.3.4
Django==3.2
django-filter==2.4.0
djangorestframework==3.12.4
logging==0.4.9.6
Markdown==3.3.4
pytz==2021.1
shortuuid==1.0.1
sqlparse==0.4.1
